Fix runtime function which was not a valid compilation target.
