+++ examples/api-gateway-authorizer/runtime/user.fn.ts
@@ -1,13 +1,17 @@
 import type { LambdaConfig } from "@notation/aws/lambda.fn";
 import { handle, json } from "@notation/aws/lambda.fn";
-import { EventWithJWTToken } from "@notation/aws/shared";
+import {
+  JWTAuthorizedApiGatewayHandler,
+  EventWithJWTToken,
+} from "@notation/aws/shared";
 import { Context } from "aws-lambda";
 
-export const getUserHandler = handle.jwtAuthorizedApiRequest(
-  (event: EventWithJWTToken, context: Context) => {
-    return json({ userId: "" });
-  },
-);
+export const getUserHandler: JWTAuthorizedApiGatewayHandler =
+  handle.jwtAuthorizedApiRequest(
+    (event: EventWithJWTToken, context: Context) => {
+      return json({ userId: "" });
+    },
+  );
 
 export const config: LambdaConfig = {
   service: "aws/lambda",
