+++ packages/aws/src/lambda.fn/handlers.ts
@@ -1,3 +1,5 @@
+import { decomposeUnverifiedJwt } from "aws-jwt-verify/jwt";
+import { APIGatewayProxyEventV2, Context } from "aws-lambda";
 import type {
   ApiGatewayHandler,
   DynamoDbBatchHandler,
@@ -14,10 +16,9 @@ export const handle = {
     async (...args) =>
       handler(...args),
 
-  jwtAuthorizedApiRequest:
-    (handler: JWTAuthorizedApiGatewayHandler): JWTAuthorizedApiGatewayHandler =>
-    async (...args) =>
-      handler(...args),
+  jwtAuthorizedApiRequest: (
+    handler: JWTAuthorizedApiGatewayHandler,
+  ): ApiGatewayHandler => toApiGatewayHandler(handler),
 
   eventBridgeScheduledEvent:
     (
@@ -41,4 +42,28 @@ export const handle = {
     (handler: SqsBatchHandler): SqsBatchHandler =>
     async (...args) =>
       handler(...args),
+};
+
+const toApiGatewayHandler = (
+  handler: JWTAuthorizedApiGatewayHandler,
+): ApiGatewayHandler => {
+  const apiGatewayHandler: ApiGatewayHandler = (
+    event: APIGatewayProxyEventV2,
+    context: Context,
+  ) => {
+    const authorizationHeader = event.headers.Authorization!;
+
+    // Remove the 'Bearer ' prefix that preceeds the token itself
+    const jwtToken = authorizationHeader?.slice(7);
+    const jwt = decomposeUnverifiedJwt(jwtToken);
+
+    const eventWithJwt = {
+      token: jwt.payload,
+      event: event,
+    };
+
+    return handler(eventWithJwt, context);
+  };
+
+  return apiGatewayHandler;
 };
